+++ frontend/app/(main)/results/ExecutiveBriefing.tsx
@@ -32,6 +32,7 @@ import { CitationContextPanel } from "@/components/synthesis/CitationContextPane
 
 interface ExecutiveBriefingProps {
   projectId: string;
+  isPublic?: boolean;
   briefing: string;
   structuredBriefing?: StructuredBriefing;
   citationMap?: Record<string, CitationInfo>;
@@ -936,6 +937,7 @@ function ReferencesList({
 
 export function ExecutiveBriefing({ 
   projectId,
+  isPublic = false,
   briefing, 
   structuredBriefing,
   citationMap, 
@@ -1517,6 +1519,7 @@ export function ExecutiveBriefing({
       <CitationContextPanel
         isOpen={Boolean(inspectedCitation)}
         projectId={projectId}
+        isPublic={isPublic}
         citationInfo={inspectedCitation?.citationInfo ?? null}
         chunkId={inspectedCitation?.chunkId}
         supportingQuote={inspectedCitation?.quote}
